+++ entities/enums.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"code.vegaprotocol.io/protos/vega"
+	v1 "code.vegaprotocol.io/protos/vega/oracles/v1"
 	"github.com/jackc/pgtype"
 )
 
@@ -434,5 +435,30 @@ func (s *VoteValue) DecodeText(_ *pgtype.ConnInfo, src []byte) error {
 		return fmt.Errorf("unknown vote value: %s", src)
 	}
 	*s = VoteValue(val)
+	return nil
+}
+
+type OracleSpecStatus v1.OracleSpec_Status
+
+const (
+	OracleSpecUnspecified = OracleSpecStatus(v1.OracleSpec_STATUS_UNSPECIFIED)
+	OracleSpecActive      = OracleSpecStatus(v1.OracleSpec_STATUS_ACTIVE)
+	OracleSpecDeactivated = OracleSpecStatus(v1.OracleSpec_STATUS_DEACTIVATED)
+)
+
+func (s OracleSpecStatus) EncodeText(_ *pgtype.ConnInfo, buf []byte) ([]byte, error) {
+	status, ok := v1.OracleSpec_Status_name[int32(s)]
+	if !ok {
+		return buf, fmt.Errorf("unknown oracle spec value: %v", s)
+	}
+	return append(buf, []byte(status)...), nil
+}
+
+func (s *OracleSpecStatus) DecodeText(_ *pgtype.ConnInfo, src []byte) error {
+	val, ok := v1.OracleSpec_Status_value[string(src)]
+	if !ok {
+		return fmt.Errorf("unknown oracle spec status: %s", src)
+	}
+	*s = OracleSpecStatus(val)
 	return nil
 }