+++ integration/integration_test.go
@@ -83,16 +83,8 @@ func waitForSIGTERM() {
 	}
 }
 
-func assertGraphQLQueriesReturnSame(t *testing.T, query string, oldResp, newResp interface{}) {
+func compareResponses(t *testing.T, oldResp, newResp interface{}) {
 	t.Helper()
-	req := graphql.NewRequest(query)
-
-	err := oldClient.Run(context.Background(), req, &oldResp)
-	require.NoError(t, err)
-
-	err = newClient.Run(context.Background(), req, &newResp)
-	require.NoError(t, err)
-
 	sortAccounts := cmpopts.SortSlices(func(a Account, b Account) bool {
 		if a.Type != b.Type {
 			return a.Type < b.Type
@@ -110,8 +102,33 @@ func assertGraphQLQueriesReturnSame(t *testing.T, query string, oldResp, newResp
 	sortVotes := cmpopts.SortSlices(func(a Vote, b Vote) bool { return a.Party.Id < b.Party.Id })
 	sortProposals := cmpopts.SortSlices(func(a Proposal, b Proposal) bool { return a.Id < b.Id })
 	sortNetParams := cmpopts.SortSlices(func(a NetworkParameter, b NetworkParameter) bool { return a.Key < b.Key })
-	diff := cmp.Diff(oldResp, newResp, sortTrades, sortVotes, sortAccounts, sortMarkets, sortProposals, sortNetParams)
+	sortSpecs := cmpopts.SortSlices(func(a, b OracleSpec) bool { return a.ID < b.ID })
+
+	diff := cmp.Diff(oldResp, newResp, sortTrades, sortVotes, sortAccounts, sortMarkets, sortProposals, sortNetParams, sortSpecs)
+
 	assert.Empty(t, diff)
+}
+
+func assertGraphQLQueriesReturnSame(t *testing.T, query string, oldResp, newResp interface{}) {
+	t.Helper()
+	req := graphql.NewRequest(query)
+
+	err := oldClient.Run(context.Background(), req, &oldResp)
+	require.NoError(t, err)
+
+	err = newClient.Run(context.Background(), req, &newResp)
+	require.NoError(t, err)
+	compareResponses(t, oldResp, newResp)
+}
+
+func assertGraphQLQueriesReturnSameIgnoreErrors(t *testing.T, query string, oldResp, newResp interface{}) {
+	t.Helper()
+	req := graphql.NewRequest(query)
+
+	_ = oldClient.Run(context.Background(), req, &oldResp)
+	_ = newClient.Run(context.Background(), req, &newResp)
+
+	compareResponses(t, oldResp, newResp)
 }
 
 func newTestConfig() (*config.Config, error) {