+++ sqlstore/migrations/0001_initial.sql
@@ -350,7 +350,7 @@ create table if not exists margin_levels (
 
 create table if not exists risk_factors (
     market_id bytea not null,
-    short numeric(32, 16) not null, 
+    short numeric(32, 16) not null,
     long numeric(32, 16) not null,
     vega_time timestamp with time zone not null references blocks(vega_time),
     primary key (market_id, vega_time)
@@ -371,6 +371,29 @@ CREATE TABLE checkpoints(
     PRIMARY KEY (block_height)
 );
 
+create type oracle_spec_status as enum('STATUS_UNSPECIFIED', 'STATUS_ACTIVE', 'STATUS_DEACTIVATED');
+
+create table if not exists oracle_specs (
+    id bytea not null,
+    created_at timestamp with time zone not null,
+    updated_at timestamp with time zone not null,
+    public_keys bytea[],
+    filters jsonb,
+    status oracle_spec_status not null,
+    vega_time timestamp with time zone not null references blocks(vega_time),
+    primary key (id, vega_time)
+);
+
+create table if not exists oracle_data (
+    public_keys bytea[],
+    data jsonb not null,
+    matched_spec_ids bytea[],
+    broadcast_at timestamp with time zone not null,
+    vega_time timestamp with time zone not null references blocks(vega_time)
+);
+
+create index if not exists idx_oracle_data_matched_spec_ids on oracle_data(matched_spec_ids);
+
 -- +goose Down
 DROP AGGREGATE IF EXISTS public.first(anyelement);
 DROP AGGREGATE IF EXISTS public.last(anyelement);
@@ -380,6 +403,11 @@ DROP FUNCTION IF EXISTS public.last_agg(anyelement, anyelement);
 DROP TABLE IF EXISTS checkpoints;
 
 DROP TABLE IF EXISTS network_parameters;
+
+DROP INDEX IF EXISTS idx_oracle_data_matched_spec_ids;
+DROP TABLE IF EXISTS oracle_data;
+DROP TABLE IF EXISTS oracle_specs;
+DROP TYPE IF EXISTS oracle_spec_status;
 
 DROP VIEW IF EXISTS votes_current;
 DROP TABLE IF EXISTS votes;